RIO.HashMap.Partial

Description

Strict HashMap partial functions. Import as:

import qualified RIO.HashMap.Partial as HM'

Basic interface

(!) :: (Eq k, Hashable k) => HashMap k v -> k -> v infixl 9

O(log n) Return the value to which the specified key is mapped. Calls error if this map contains no mapping for the key.