RIO.Time

Documentation

formatTime :: FormatTime t => TimeLocale -> String -> t -> String

Substitute various time-related information for each %-code in the string, as per formatCharacter.

The general form is %<modifier><width><specifier>, where <modifier> and <width> are optional.

<modifier>

glibc-style modifiers can be used before the specifier (here marked as z):

%-z

no padding

%_z

pad with spaces

%0z

pad with zeros

%^z

convert to upper case

%#z

convert to lower case (consistently, unlike glibc)

<width>

Width digits can also be used after any modifiers and before the specifier (here marked as z), for example:

%4z

pad to 4 characters (with default padding character)

%_12z

pad with spaces to 12 characters

<specifier>

For all types (note these three are done by formatTime, not by formatCharacter):

%%

%

%t

tab

%n

newline

TimeZone

For TimeZone (and ZonedTime and UTCTime):

%z

timezone offset in the format -HHMM.

%Z

timezone name

LocalTime

For LocalTime (and ZonedTime and UTCTime and UniversalTime):

%c

as dateTimeFmt locale (e.g. %a %b %e %H:%M:%S %Z %Y)

TimeOfDay

For TimeOfDay (and LocalTime and ZonedTime and UTCTime and UniversalTime):

%R

same as %H:%M

%T

same as %H:%M:%S

%X

as timeFmt locale (e.g. %H:%M:%S)

%r

as time12Fmt locale (e.g. %I:%M:%S %p)

%P

day-half of day from (amPm locale), converted to lowercase, am, pm

%p

day-half of day from (amPm locale), AM, PM

%H

hour of day (24-hour), 0-padded to two chars, 00 - 23

%k

hour of day (24-hour), space-padded to two chars, 0 - 23

%I

hour of day-half (12-hour), 0-padded to two chars, 01 - 12

%l

hour of day-half (12-hour), space-padded to two chars, 1 - 12

%M

minute of hour, 0-padded to two chars, 00 - 59

%S

second of minute (without decimal part), 0-padded to two chars, 00 - 60

%q

picosecond of second, 0-padded to twelve chars, 000000000000 - 999999999999.

%Q

decimal point and fraction of second, up to 12 second decimals, without trailing zeros. For a whole number of seconds, %Q omits the decimal point unless padding is specified.

UTCTime and ZonedTime

For UTCTime and ZonedTime:

%s

number of whole seconds since the Unix epoch. For times before the Unix epoch, this is a negative number. Note that in %s.%q and %s%Q the decimals are positive, not negative. For example, 0.9 seconds before the Unix epoch is formatted as -1.1 with %s%Q.

Day

For Day (and LocalTime and ZonedTime and UTCTime and UniversalTime):

%D

same as %m/%d/%y

%F

same as %Y-%m-%d

%x

as dateFmt locale (e.g. %m/%d/%y)

%Y

year, no padding. Note %0Y and %_Y pad to four chars

%y

year of century, 0-padded to two chars, 00 - 99

%C

century, no padding. Note %0C and %_C pad to two chars

%B

month name, long form (fst from months locale), January - December

%b, %h

month name, short form (snd from months locale), Jan - Dec

%m

month of year, 0-padded to two chars, 01 - 12

%d

day of month, 0-padded to two chars, 01 - 31

%e

day of month, space-padded to two chars, 1 - 31

%j

day of year, 0-padded to three chars, 001 - 366

%f

century for Week Date format, no padding. Note %0f and %_f pad to two chars

%V

week of year for Week Date format, 0-padded to two chars, 01 - 53

%u

day of week for Week Date format, 1 - 7

%a

day of week, short form (snd from wDays locale), Sun - Sat

%A

day of week, long form (fst from wDays locale), Sunday - Saturday

%U

week of year where weeks start on Sunday (as sundayStartWeek), 0-padded to two chars, 00 - 53

%w

day of week number, 0 (= Sunday) - 6 (= Saturday)

%W

week of year where weeks start on Monday (as mondayStartWeek), 0-padded to two chars, 00 - 53

type NumericPadOption = Maybe Char

class FormatTime t where

Methods

formatCharacter :: Char -> Maybe (TimeLocale -> Maybe NumericPadOption -> Maybe Int -> t -> String)

readsTime

Arguments

:: ParseTime t

=> TimeLocale
Time locale.

-> String
Format string

-> ReadS t

readTime

Arguments

:: ParseTime t

=> TimeLocale
Time locale.

-> String
Format string.

-> String
Input string.

-> t
The time value.

parseTime

Arguments

:: ParseTime t

=> TimeLocale
Time locale.

-> String
Format string.

-> String
Input string.

-> Maybe t
The time value, or Nothing if the input could not be parsed using the given format.

readPTime

Arguments

:: ParseTime t

=> Bool
Accept leading whitespace?

-> TimeLocale
Time locale.

-> String
Format string

-> ReadP t

Parse a time value given a format string. See parseTimeM for details.

readSTime

Arguments

:: ParseTime t

=> Bool
Accept leading whitespace?

-> TimeLocale
Time locale.

-> String
Format string

-> ReadS t

Parse a time value given a format string. See parseTimeM for details.

parseTimeOrError

Arguments

:: ParseTime t

=> Bool
Accept leading and trailing whitespace?

-> TimeLocale
Time locale.

-> String
Format string.

-> String
Input string.

-> t
The time value.

Parse a time value given a format string. Fails if the input could not be parsed using the given format. See parseTimeM for details.

parseTimeM

Arguments

:: (Monad m, ParseTime t)

=> Bool
Accept leading and trailing whitespace?

-> TimeLocale
Time locale.

-> String
Format string.

-> String
Input string.

-> m t
Return the time value, or fail if the input could not be parsed using the given format.

Parses a time value given a format string. Supports the same %-codes as formatTime, including %-, %_ and %0 modifiers, however padding widths are not supported. Case is not significant in the input string. Some variations in the input are accepted:

%z

accepts any of -HHMM or -HH:MM.

%Z

accepts any string of letters, or any of the formats accepted by %z.

%0Y

accepts exactly four digits.

%0G

accepts exactly four digits.

%0C

accepts exactly two digits.

%0f

accepts exactly two digits.

class ParseTime t where

The class of types which can be parsed given a UNIX-style time format string.

Methods

buildTime

Arguments

:: TimeLocale
The time locale.

-> [(Char, String)]
Pairs of format characters and the corresponding part of the input.

-> Maybe t

Builds a time value from a parsed input string. If the input does not include all the information needed to construct a complete value, any missing parts should be taken from 1970-01-01 00:00:00 +0000 (which was a Thursday). In the absence of %C or %Y, century is 1969 - 2068.

zonedTimeToUTC :: ZonedTime -> UTCTime

utcToZonedTime :: TimeZone -> UTCTime -> ZonedTime

data ZonedTime

A local time together with a time zone.

Constructors

ZonedTime

Fields

zonedTimeToLocalTime :: LocalTime

zonedTimeZone :: TimeZone

rfc822DateFormat :: String

Format string according to RFC822.

iso8601DateFormat :: Maybe String -> String

Construct format string according to ISO-8601.

The Maybe String argument allows to supply an optional time specification. E.g.:

iso8601DateFormat Nothing            == "%Y-%m-%d"           -- i.e. YYYY-MM-DD
iso8601DateFormat (Just "%H:%M:%S")  == "%Y-%m-%dT%H:%M:%S"  -- i.e. YYYY-MM-DDTHH:MM:SS

defaultTimeLocale :: TimeLocale

Locale representing American usage.

knownTimeZones contains only the ten time-zones mentioned in RFC 822 sec. 5: "UT", "GMT", "EST", "EDT", "CST", "CDT", "MST", "MDT", "PST", "PDT". Note that the parsing functions will regardless parse single-letter military time-zones and +HHMM format.

data TimeLocale

Constructors

TimeLocale

Fields

wDays :: [(String, String)]

full and abbreviated week days, starting with Sunday

months :: [(String, String)]

full and abbreviated months

amPm :: (String, String)

AM/PM symbols

dateTimeFmt :: String

formatting strings

dateFmt :: String

formatting strings

timeFmt :: String

formatting strings

time12Fmt :: String

formatting strings

knownTimeZones :: [TimeZone]

time zones known by name

localTimeToUT1 :: Rational -> LocalTime -> UniversalTime

Get the UT1 time of a local time on a particular meridian (in degrees, positive is East).

ut1ToLocalTime :: Rational -> UniversalTime -> LocalTime

Get the local time of a UT1 time on a particular meridian (in degrees, positive is East).

localTimeToUTC :: TimeZone -> LocalTime -> UTCTime

Get the UTC time of a local time in a time zone.

utcToLocalTime :: TimeZone -> UTCTime -> LocalTime

Get the local time of a UTC time in a time zone.

data LocalTime

A simple day and time aggregate, where the day is of the specified parameter, and the time is a TimeOfDay. Conversion of this (as local civil time) to UTC depends on the time zone. Conversion of this (as local mean time) to UT1 depends on the longitude.

Constructors

LocalTime

Fields

localDay :: Day

localTimeOfDay :: TimeOfDay

timeOfDayToDayFraction :: TimeOfDay -> Rational

Get the fraction of a day since midnight given a time of day.

dayFractionToTimeOfDay :: Rational -> TimeOfDay

Get the time of day given the fraction of a day since midnight.

timeOfDayToTime :: TimeOfDay -> DiffTime

Get the time since midnight for a given time of day.

timeToTimeOfDay :: DiffTime -> TimeOfDay

Get the time of day given a time since midnight. Time more than 24h will be converted to leap-seconds.

localToUTCTimeOfDay :: TimeZone -> TimeOfDay -> (Integer, TimeOfDay)

Convert a time of day in some timezone to a time of day in UTC, together with a day adjustment.

utcToLocalTimeOfDay :: TimeZone -> TimeOfDay -> (Integer, TimeOfDay)

Convert a time of day in UTC to a time of day in some timezone, together with a day adjustment.

makeTimeOfDayValid :: Int -> Int -> Pico -> Maybe TimeOfDay

midday :: TimeOfDay

Hour twelve

midnight :: TimeOfDay

Hour zero

data TimeOfDay

Time of day as represented in hour, minute and second (with picoseconds), typically used to express local time of day.

Constructors

TimeOfDay

Fields

todHour :: Int

range 0 - 23

todMin :: Int

range 0 - 59

todSec :: Pico

Note that 0 <= todSec < 61, accomodating leap seconds. Any local minute may have a leap second, since leap seconds happen in all zones simultaneously

utc :: TimeZone

The UTC time zone.

timeZoneOffsetString :: TimeZone -> String

Text representing the offset of this timezone, such as "-0800" or "+0400" (like %z in formatTime).

timeZoneOffsetString' :: Maybe Char -> TimeZone -> String

Text representing the offset of this timezone, such as "-0800" or "+0400" (like %z in formatTime), with arbitrary padding.

hoursToTimeZone :: Int -> TimeZone

Create a nameless non-summer timezone for this number of hours.

minutesToTimeZone :: Int -> TimeZone

Create a nameless non-summer timezone for this number of minutes.

data TimeZone

A TimeZone is a whole number of minutes offset from UTC, together with a name and a "just for summer" flag.

Constructors

TimeZone

Fields

timeZoneMinutes :: Int

The number of minutes offset from UTC. Positive means local time will be later in the day than UTC.

timeZoneSummerOnly :: Bool

Is this time zone just persisting for the summer?

timeZoneName :: String

The name of the zone, typically a three- or four-letter acronym.

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

newtype UniversalTime

The Modified Julian Date is the day with the fraction of the day, measured from UT midnight. It's used to represent UT1, which is time as measured by the earth's rotation, adjusted for various wobbles.

Constructors

ModJulianDate

Fields

getModJulianDate :: Rational

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)

getTime_resolution :: DiffTime

The resolution of getSystemTime, getCurrentTime, getPOSIXTime

nominalDay :: NominalDiffTime

One day in NominalDiffTime.

data NominalDiffTime

This is a length of time, as measured by UTC. Conversion functions will treat it as seconds. It has a precision of 10^-12 s. It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

diffTimeToPicoseconds :: DiffTime -> Integer

Get the number of picoseconds in a DiffTime.

picosecondsToDiffTime :: Integer -> DiffTime

Create a DiffTime from a number of picoseconds.

secondsToDiffTime :: Integer -> DiffTime

Create a DiffTime which represents an integral number of seconds.

data DiffTime

This is a length of time, as measured by a clock. Conversion functions will treat it as seconds. It has a precision of 10^-12 s.

addGregorianYearsRollOver :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th rolled over to Mar 1st if necessary. For instance, 2004-02-29 + 2 years = 2006-03-01.

addGregorianYearsClip :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th clipped to Feb 28th if necessary. For instance, 2004-02-29 + 2 years = 2006-02-28.

addGregorianMonthsRollOver :: Integer -> Day -> Day

Add months, with days past the last day of the month rolling over to the next month. For instance, 2005-01-30 + 1 month = 2005-03-02.

addGregorianMonthsClip :: Integer -> Day -> Day

Add months, with days past the last day of the month clipped to the last day. For instance, 2005-01-30 + 1 month = 2005-02-28.

gregorianMonthLength :: Integer -> Int -> Int

The number of days in a given month according to the proleptic Gregorian calendar. First argument is year, second is month.

showGregorian :: Day -> String

Show in ISO 8601 format (yyyy-mm-dd)

fromGregorianValid :: Integer -> Int -> Int -> Maybe Day

Convert from proleptic Gregorian calendar. First argument is year, second month number (1-12), third day (1-31). Invalid values will return Nothing

fromGregorian :: Integer -> Int -> Int -> Day

Convert from proleptic Gregorian calendar. First argument is year, second month number (1-12), third day (1-31). Invalid values will be clipped to the correct range, month first, then day.

toGregorian :: Day -> (Integer, Int, Int)

Convert to proleptic Gregorian calendar. First element of result is year, second month number (1-12), third day (1-31).

isLeapYear :: Integer -> Bool

Is this year a leap year according to the proleptic Gregorian calendar?

diffDays :: Day -> Day -> Integer

addDays :: Integer -> Day -> Day

newtype Day

The Modified Julian Day is a standard count of days, with zero being the day 1858-11-17.

Constructors

ModifiedJulianDay

Fields

toModifiedJulianDay :: Integer

getCurrentTime :: MonadIO m => m UTCTime

getTimeZone :: MonadIO m => UTCTime -> m TimeZone

getCurrentTimeZone :: MonadIO m => m TimeZone

getZonedTime :: MonadIO m => m ZonedTime

utcToLocalZonedTime :: MonadIO m => UTCTime -> m ZonedTime