RIO.Vector.Unboxed

Description

Unboxed Vector. Import as:

import qualified RIO.Vector.Unboxed as VU

This module does not export any partial or unsafe functions. For those, see RIO.Vector.Unboxed.Partial and RIO.Vector.Unboxed.Unsafe

Unboxed vectors

data family Vector a

data family MVector s a

class (Vector Vector a, MVector MVector a) => Unbox a

Accessors

Length information

length :: Unbox a => Vector a -> Int

O(1) Yield the length of the vector.

null :: Unbox a => Vector a -> Bool

O(1) Test whether a vector is empty.

Indexing

(!?) :: Unbox a => Vector a -> Int -> Maybe a

O(1) Safe indexing.

Extracting subvectors

slice

Arguments

:: Unbox a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

take :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements, in which case it is returned unchanged.

drop :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements, in which case an empty vector is returned.

splitAt :: Unbox a => Int -> Vector a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

Since: vector-0.7.1

Construction

Initialisation

empty :: Unbox a => Vector a

O(1) The empty vector.

singleton :: Unbox a => a -> Vector a

O(1) A vector with exactly one element.

replicate :: Unbox a => Int -> a -> Vector a

O(n) A vector of the given length with the same value in each position.

generate :: Unbox a => Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index.

iterateN :: Unbox a => Int -> (a -> a) -> a -> Vector a

O(n) Apply the function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

\( \underbrace{x, f (x), f (f (x)), \ldots}_{\max(0,n)\rm{~elements}} \)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.iterateN 0 undefined undefined :: VU.Vector Int
[]
>>> VU.iterateN 3 (\(i, c) -> (pred i, succ c)) (0 :: Int, 'a')
[(0,'a'),(-1,'b'),(-2,'c')]

Since: vector-0.7.1

Monadic initialisation

replicateM :: (Monad m, Unbox a) => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Unbox a) => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index.

iterateNM :: (Monad m, Unbox a) => Int -> (a -> m a) -> a -> m (Vector a)

O(n) Apply the monadic function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

For a non-monadic version, see iterateN.

Since: vector-0.12.0.0

create :: Unbox a => (forall s. ST s (MVector s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: (Traversable f, Unbox a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: Unbox a => (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Unbox a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrM :: (Monad m, Unbox a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: (Monad m, Unbox a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

constructN :: Unbox a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in <a,b,c>

constructrN :: Unbox a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in <c,b,a>

Enumeration

enumFromN :: (Unbox a, Num a) => a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: (Unbox a, Num a) => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 2 5 = <1,3,5,7,9>

enumFromTo :: (Unbox a, Enum a) => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If possible, use enumFromN instead.

enumFromThenTo :: (Unbox a, Enum a) => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If possible, use enumFromStepN instead.

Concatenation

cons :: Unbox a => a -> Vector a -> Vector a

O(n) Prepend an element.

snoc :: Unbox a => Vector a -> a -> Vector a

O(n) Append an element.

(++) :: Unbox a => Vector a -> Vector a -> Vector a infixr 5

O(m+n) Concatenate two vectors.

concat :: Unbox a => [Vector a] -> Vector a

O(n) Concatenate all vectors in the list.

Restricting memory usage

force :: Unbox a => Vector a -> Vector a

O(n) Yield the argument, but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Permutations

reverse :: Unbox a => Vector a -> Vector a

O(n) Reverse a vector.

Safe destructive update

modify :: Unbox a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Apply a destructive operation to a vector. The operation may be performed in place if it is safe to do so and will modify a copy of the vector otherwise (see New for details).

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> import qualified Data.Vector.Unboxed.Mutable as MVU
>>> VU.modify (\v -> MVU.write v 0 'x') $ VU.replicate 4 'a'
"xaaa"

Elementwise operations

Indexing

indexed :: Unbox a => Vector a -> Vector (Int, a)

O(n) Pair each element in a vector with its index.

Mapping

map :: (Unbox a, Unbox b) => (a -> b) -> Vector a -> Vector b

O(n) Map a function over a vector.

imap :: (Unbox a, Unbox b) => (Int -> a -> b) -> Vector a -> Vector b

O(n) Apply a function to every element of a vector and its index.

concatMap :: (Unbox a, Unbox b) => (a -> Vector b) -> Vector a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Unbox a, Unbox b) => (a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results.

imapM :: (Monad m, Unbox a, Unbox b) => (Int -> a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to every element of a vector and its index, yielding a vector of results.

mapM_ :: (Monad m, Unbox a) => (a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results.

imapM_ :: (Monad m, Unbox a) => (Int -> a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to every element of a vector and its index, ignoring the results.

forM :: (Monad m, Unbox a, Unbox b) => Vector a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: (Monad m, Unbox a) => Vector a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

Zipping

zipWith :: (Unbox a, Unbox b, Unbox c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

zip :: (Unbox a, Unbox b) => Vector a -> Vector b -> Vector (a, b)

O(1) Zip 2 vectors.

zip3 :: (Unbox a, Unbox b, Unbox c) => Vector a -> Vector b -> Vector c -> Vector (a, b, c)

O(1) Zip 3 vectors.

zip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector a -> Vector b -> Vector c -> Vector d -> Vector (a, b, c, d)

O(1) Zip 4 vectors.

zip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector (a, b, c, d, e)

O(1) Zip 5 vectors.

zip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector (a, b, c, d, e, f)

O(1) Zip 6 vectors.

Monadic zipping

zipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results.

izipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and yield a vector of results.

zipWithM_ :: (Monad m, Unbox a, Unbox b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results.

izipWithM_ :: (Monad m, Unbox a, Unbox b) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and ignore the results.

Unzipping

unzip :: (Unbox a, Unbox b) => Vector (a, b) -> (Vector a, Vector b)

O(1) Unzip 2 vectors.

unzip3 :: (Unbox a, Unbox b, Unbox c) => Vector (a, b, c) -> (Vector a, Vector b, Vector c)

O(1) Unzip 3 vectors.

unzip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector (a, b, c, d) -> (Vector a, Vector b, Vector c, Vector d)

O(1) Unzip 4 vectors.

unzip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector (a, b, c, d, e) -> (Vector a, Vector b, Vector c, Vector d, Vector e)

O(1) Unzip 5 vectors.

unzip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector (a, b, c, d, e, f) -> (Vector a, Vector b, Vector c, Vector d, Vector e, Vector f)

O(1) Unzip 6 vectors.

Working with predicates

Filtering

filter :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate.

ifilter :: Unbox a => (Int -> a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate which is applied to the values and their indices.

uniq :: (Unbox a, Eq a) => Vector a -> Vector a

O(n) Drop repeated adjacent elements. The first element in each group is returned.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.uniq $ VU.fromList [1,3,3,200,3 :: Int]
[1,3,200,3]
>>> import Data.Semigroup
>>> VU.uniq $ VU.fromList [ Arg 1 'a', Arg 1 'b', Arg (1 :: Int) 'c']
[Arg 1 'a']

mapMaybe :: (Unbox a, Unbox b) => (a -> Maybe b) -> Vector a -> Vector b

O(n) Map the values and collect the Just results.

imapMaybe :: (Unbox a, Unbox b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

O(n) Map the indices/values and collect the Just results.

filterM :: (Monad m, Unbox a) => (a -> m Bool) -> Vector a -> m (Vector a)

O(n) Drop all elements that do not satisfy the monadic predicate.

takeWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate. The current implementation is not copy-free, unless the result vector is fused away.

dropWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved, but the operation is often faster than partition.

span :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

Searching

elem :: (Unbox a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector contains an element.

notElem :: (Unbox a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem).

find :: Unbox a => (a -> Bool) -> Vector a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Unbox a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndices :: Unbox a => (a -> Bool) -> Vector a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Unbox a, Eq a) => a -> Vector a -> Maybe Int

O(n) Yield Just the index of the first occurrence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Unbox a, Eq a) => a -> Vector a -> Vector Int

O(n) Yield the indices of all occurrences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold.

foldl' :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator.

foldr :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold.

foldr' :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with a strict accumulator.

ifoldl :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold using a function applied to each element and its index.

ifoldl' :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator using a function applied to each element and its index.

ifoldr :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold using a function applied to each element and its index.

ifoldr' :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with strict accumulator using a function applied to each element and its index.

Specialised folds

all :: Unbox a => (a -> Bool) -> Vector a -> Bool

O(n) Check if all elements satisfy the predicate.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.all even $ VU.fromList [2, 4, 12 :: Int]
True
>>> VU.all even $ VU.fromList [2, 4, 13 :: Int]
False
>>> VU.all even (VU.empty :: VU.Vector Int)
True

any :: Unbox a => (a -> Bool) -> Vector a -> Bool

O(n) Check if any element satisfies the predicate.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.any even $ VU.fromList [1, 3, 7 :: Int]
False
>>> VU.any even $ VU.fromList [3, 2, 13 :: Int]
True
>>> VU.any even (VU.empty :: VU.Vector Int)
False

and :: Vector Bool -> Bool

O(n) Check if all elements are True.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.and $ VU.fromList [True, False]
False
>>> VU.and VU.empty
True

or :: Vector Bool -> Bool

O(n) Check if any element is True.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.or $ VU.fromList [True, False]
True
>>> VU.or VU.empty
False

sum :: (Unbox a, Num a) => Vector a -> a

O(n) Compute the sum of the elements.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.sum $ VU.fromList [300,20,1 :: Int]
321
>>> VU.sum (VU.empty :: VU.Vector Int)
0

product :: (Unbox a, Num a) => Vector a -> a

O(n) Compute the product of the elements.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.product $ VU.fromList [1,2,3,4 :: Int]
24
>>> VU.product (VU.empty :: VU.Vector Int)
1

Monadic folds

foldM :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold.

ifoldM :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold using a function applied to each element and its index.

foldM' :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator.

ifoldM' :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

foldM_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result.

ifoldM_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result using a function applied to each element and its index.

foldM'_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result.

ifoldM'_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result using a function applied to each element and its index.

Prefix sums (scans)

prescanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan.

prescanl f z = init . scanl f z

Examples

Expand

>>> import qualified Data.Vector as VU
>>> VU.prescanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[0,1,3,6]

prescanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan with strict accumulator.

postscanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan.

postscanl f z = tail . scanl f z

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.postscanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[1,3,6,10]

postscanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan with strict accumulator.

scanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan.

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[0,1,3,6,10]

scanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan with strict accumulator.

prescanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan.

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan with strict accumulator.

postscanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan.

postscanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan with strict accumulator.

scanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan.

scanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan with strict accumulator.

Conversions

Lists

toList :: Unbox a => Vector a -> [a]

O(n) Convert a vector to a list.

fromList :: Unbox a => [a] -> Vector a

O(n) Convert a list to a vector.

fromListN :: Unbox a => Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector. It's expected that the supplied list will be exactly n elements long. As an optimization, this function allocates a buffer for n elements, which could be used for DoS-attacks by exhausting the memory if an attacker controls that parameter.

fromListN n xs = fromList (take n xs)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.fromListN 3 [1,2,3,4,5 :: Int]
[1,2,3]
>>> VU.fromListN 3 [1 :: Int]
[1]

Different vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert between different vector types.

Mutable vectors

freeze :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (Unbox a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of an immutable vector.

copy :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.